OnelineFetch: Move from utils to own module
Required since setup requires uninstalled modules

Fixes https://github.com/NIT-dgp/UIP/issues/216
